int -> long
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/branches/PHP_5_3@261790 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/standard/string.c
@@ -998,7 +998,7 @@ PHPAPI void php_explode_negative_limit(zval *delim, zval *str, zval *return_valu
 PHP_FUNCTION(explode)
 {
 	zval **str, **delim, **zlimit = NULL;
-	int limit = -1;
+	long limit = -1;
 	int argc = ZEND_NUM_ARGS();
 	
 	if (zend_parse_parameters(ZEND_NUM_ARGS() TSRMLS_CC, "ZZ|l", &delim, &str, &limit) == FAILURE) {
